First revision of script for creating limited ssh user
Often there's a need for a user that can be used for establishing a
remote ssh tunnel, but that also has 0 priviledges in the event that the
keys should be lost. This script creates that user and overwrites the
sshd_config file, to make sure everything is set up correctly. Do keep 2
things in mind hovewer: 1. this script is only supposed to be run on a
fresh arch installation and 2. there is still some warning to implement
and a lot of testing before it it safe to run.